move daily report fields
